(ver3.5) adding ability to store and show past bills

+++ app/models.py
@@ -507,7 +507,7 @@ class billingHistoryTable:
 
                 CREATE TABLE IF NOT EXISTS billingHistory
                     (
-                        id INTEGER NOT NULL,
+                        id BIGINT NOT NULL,
                         bill TEXT NOT NULL,
                         billDate TEXT NOT NULL
                     )
